83d7fd8ce47caaf5d8383e2fd2e7615a5e9ff7e1

+++ ejaf_maintenance_equipment/models/maintenance_equipment.py
@@ -58,7 +58,7 @@ class MaintenanceEquipment(models.Model):
     _inherit = 'maintenance.equipment'
 
     def _get_reservation_liters(self):
-        reservation_liters = int(self.env['ir.config_parameter'].sudo().get_param('reservation_liters'))
+        reservation_liters = float(self.env['ir.config_parameter'].sudo().get_param('reservation_liters'))
         return reservation_liters
 
     site = fields.Char(string="Code", tracking=True)
